# Optimize CI visual regression snapshotting