--- a/scripts/snapshot.js
+++ b/scripts/snapshot.js
@@ -25,17 +25,19 @@ async function getScreenshot(page, url) {
 const esbuildContext = await esbuild.context({});
 const { port } = await esbuildContext.serve({ servedir: 'dist' });
 const browser = await puppeteer.launch({ headless: 'new' });
+const browserContext = await browser.createIncognitoBrowserContext();
 const localURL = `http://localhost:${port}/`;
 const outputDirectory = join('tmp/screenshot/branches', branch);
-const page = await browser.newPage();
 
 await mkdir(outputDirectory, { recursive: true });
-for await (const path of paths) {
-  const screenshot = await getScreenshot(page, localURL + path);
-  const filename = join(outputDirectory, `${basename(path, extname(path))}.png`);
-  process.stdout.write(`Writing ${filename}...\n`);
-  await writeFile(filename, screenshot);
-}
-
-browser.close();
-esbuildContext.dispose();
+await Promise.all(
+  paths.map(async (path) => {
+    const page = await browserContext.newPage();
+    const screenshot = await getScreenshot(page, localURL + path);
+    const filename = join(outputDirectory, `${basename(path, extname(path))}.png`);
+    process.stdout.write(`Writing ${filename}...\n`);
+    await writeFile(filename, screenshot);
+    await page.close();
+  }),
+);
+await Promise.all([browserContext.close(), browser.close(), esbuildContext.dispose()]);