@aduth 2024-02-09
Noting for future reference that running the snapshotting with a single page in series (70d3a7eab0b7430900f5fa7cad6ad2c2ebf3bf51) seemed to be quite performant as well. I expect that the parallelization starts hitting some bottlenecks when trying to create so many pages in parallel, and there's some benefit in limiting the concurrency, maybe with one or a few dedicated "page" instances, and maybe with a library that could help simplify this (e.g. [`p-map`](https://www.npmjs.com/package/p-map)). I didn't go too far down this rabbit-hole since I've started hitting diminishing returns with these changes, but something to look at in the future if performance becomes a concern.